// src/integrations/supabase/types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5";
  };
  public: {
    Tables: {
// ...
      form_submissions: {
        Row: {
          admin_notes: string | null;
          created_at: string;
          email: string | null;
          id: string;
          message: string | null;
          meta: Json | null;
          name: string;
          phone: string | null;
          status: Database["public"]["Enums"]["submission_status"];
          type: Database["public"]["Enums"]["submission_type"];
          updated_at: string;
        };
        Insert: {
          admin_notes?: string | null;
          created_at?: string;
          email?: string | null;
          id?: string;
          message?: string | null;
          meta?: Json | null;
          name: string;
          phone?: string | null;
          status?: Database["public"]["Enums"]["submission_status"];
          type: Database["public"]["Enums"]["submission_type"];
          updated_at?: string;
        };
        Update: {
          admin_notes?: string | null;
          created_at?: string;
          email?: string | null;
          id?: string;
          message?: string | null;
          meta?: Json | null;
          name?: string;
          phone?: string | null;
          status?: Database["public"]["Enums"]["submission_status"];
          type?: Database["public"]["Enums"]["submission_type"];
          updated_at?: string;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    keyof DefaultSchema["Enums"] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;